ortalab
Module joker

Structs

AbstractJoker

BaronJoker

BasicJoker

BlackboardJoker

BlueprintJoker

CleverJoker

CraftyJoker

CrazyJoker

DeviousJoker

DrollJoker

EvenStevenJoker

FibonacciJoker

FlowerPotJoker

FourFingersJoker

GluttonousJoker

GreedyJoker

JokerContext

JokerFactory

JollyJoker

LustyJoker

MadJoker

MimeJoker

OddToddJoker

PareidoliaJoker

PhotographJoker

RaisedFistJoker

ScaryFaceJoker

ShortcutJoker

SlyJoker

SmearedJoker

SmileyFaceJoker

SockAndBuskinJoker

SplashJoker

WilyJoker

WrathfulJoker

ZanyJoker

Enums

JokerActivation

ScoringScope

Traits

JokerEffect

Functions

apply_joker_effects

Applies independent joker effects to chips and mult.

apply_jokers_retrigger

Applies retriggerable joker effects for cards in hand.

apply_onheld_joker_effects

Applies OnHeld joker effects for a specific card.

apply_onscored_joker_effects

Applies OnScored joker effects for a specific card.

find_lowest_rank_card 🔒

Finds the card with the lowest rank in a set.

get_card_numeric_value 🔒

Gets the numeric value of a card’s rank.

get_joker_id

Returns a unique ID for a joker type.

has_flush 🔒

Checks if a flush is present in the cards.

has_straight 🔒

Checks if a straight is present in the cards.