ortalab
Module modifiers

Structs

BonusEnhancement

FoilEdition

GlassEnhancement

HolographicEdition

MultEnhancement

PolychromeEdition

SteelEnhancement

WildEnhancement

Traits

Edition

Enhancement

Functions

apply_edition_effect

Applies an edition effect directly to chips and mult based on the edition type.

apply_enhancements

Applies enhancements and editions to a set of cards, modifying chips and mult.

create_edition_handler

Creates an edition handler based on the edition type.

create_enhancement_handler

Creates an enhancement handler based on the enhancement type.

get_possible_straight_sequences 🔒

Generates all possible straight sequences of ranks.

handle_all_wild_cards

Converts a set of all wild cards into a straight sequence if possible.

handle_wild

Handles wild cards by adjusting the card set, potentially forming a straight.

select_best_straight_sequence 🔒

Selects the best straight sequence from possible options.

select_best_suit 🔒

Determines the most common suit among the cards.

try_form_wild_straight

Attempts to form a straight using wild cards to fill gaps.